Refactor: 모듈 방식으로 수정
