+++ index.html
@@ -10,10 +10,8 @@
       <div class="logo">
         <img src="src/assets/title.png" alt="title" />
       </div>
-      <div class="date">
-        <span id="current-date"></span>
-      </div>
+      <div class="date" id="current-date"></div>
     </header>
-    <script src="src/js/app.js"></script>
+    <script type="module" src="src/js/app.js"></script>
   </body>
 </html>
