+++ src/js/app.js
@@ -1,20 +1,6 @@
-const currentDate = new Date();
+import { getFormattedDate } from "./utils/dateUtils.js";
 
-const year = currentDate.getFullYear();
-const month = String(currentDate.getMonth() + 1).padStart(2, "0");
-const date = String(currentDate.getDate()).padStart(2, "0");
-
-const week = [
-  "일요일",
-  "월요일",
-  "화요일",
-  "수요일",
-  "목요일",
-  "금요일",
-  "토요일",
-];
-const day = week[currentDate.getDay()];
-
-const formattedDate = `${year}. ${month}. ${date}. ${day}`;
-
-document.getElementById("current-date").textContent = formattedDate;
+// DOMContentLoaded 이벤트 발생 시 날짜를 설정
+document.addEventListener("DOMContentLoaded", () => {
+  document.getElementById("current-date").textContent = getFormattedDate();
+});
